Remove Windows specific library loading (libs must be in the search path now), get user home directory now works multi-platform and relocated test utils to the correct namespace.
