+++ wrappers/dotnet/indy-sdk-dotnet-test/AnonCredsTests/AnonCredsIntegrationTestBase.cs
@@ -1,5 +1,6 @@
 ﻿using Hyperledger.Indy.Sdk.AnonCredsApi;
 using Hyperledger.Indy.Sdk.WalletApi;
+using Hyperledger.Indy.Sdk.Test.Util;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using System.Threading.Tasks;
