+++ wrappers/dotnet/indy-sdk-dotnet-test/DemoTests/AgentDemoTest.cs
@@ -3,6 +3,7 @@ using Hyperledger.Indy.Sdk.LedgerApi;
 using Hyperledger.Indy.Sdk.PoolApi;
 using Hyperledger.Indy.Sdk.SignUsApi;
 using Hyperledger.Indy.Sdk.WalletApi;
+using Hyperledger.Indy.Sdk.Test.Util;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using System.Threading.Tasks;
