+++ wrappers/dotnet/indy-sdk-dotnet-test/DemoTests/AnonCredsDemoTest.cs
@@ -1,6 +1,7 @@
 ﻿using Hyperledger.Indy.Sdk.AnonCredsApi;
 using Hyperledger.Indy.Sdk.PoolApi;
 using Hyperledger.Indy.Sdk.WalletApi;
+using Hyperledger.Indy.Sdk.Test.Util;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Newtonsoft.Json.Linq;
 using System;
