+++ wrappers/dotnet/indy-sdk-dotnet-test/LedgerTests/RequestTest.cs
@@ -2,6 +2,7 @@
 using Hyperledger.Indy.Sdk.PoolApi;
 using Hyperledger.Indy.Sdk.SignUsApi;
 using Hyperledger.Indy.Sdk.WalletApi;
+using Hyperledger.Indy.Sdk.Test.Util;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Newtonsoft.Json.Linq;
 using System.Threading.Tasks;
