+++ wrappers/dotnet/indy-sdk-dotnet-test/PoolTests/ClosePoolTest.cs
@@ -1,4 +1,5 @@
 ﻿using Hyperledger.Indy.Sdk.PoolApi;
+using Hyperledger.Indy.Sdk.Test.Util;
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using System.Threading.Tasks;
