+++ wrappers/dotnet/indy-sdk-dotnet-test/PoolTests/RefreshPoolTest.cs
@@ -1,4 +1,5 @@
-﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
+﻿using Hyperledger.Indy.Sdk.Test.Util;
+using Microsoft.VisualStudio.TestTools.UnitTesting;
 using System.Threading.Tasks;
 
 namespace Hyperledger.Indy.Sdk.Test.PoolTests
