+++ wrappers/dotnet/indy-sdk-dotnet-test/indy-sdk-dotnet-test.csproj
@@ -1,7 +1,7 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFrameworks>net45;netcoreapp1.0</TargetFrameworks>
+    <TargetFrameworks>netcoreapp2.0</TargetFrameworks>
     <AssemblyName>Hyperledger.Indy.Sdk.Test</AssemblyName>
     <RootNamespace>Hyperledger.Indy.Sdk.Test</RootNamespace>
   </PropertyGroup>
