--- wrappers/dotnet/lib/.gitignore
@@ -1,2 +1,0 @@
-*
-
